+++ CHANGELOG.md
@@ -9,6 +9,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 This shows all commits since last release (actually, currently all commits).
 
+## [2.8.1](https://github.com/wickeyware/wickeyappstore/compare/2.8.0...2.8.1) - 2018-04-23
+
+### Fixed
+
+* **UserService** Subscribe to getStore to ensure the call is made.
+* **UserService** Subscribe to setStore to ensure the call is made.
+* **UserService** Subscribe to deleteStore to ensure the call is made.
+* **ApiConnection** Remove console debug out on getStore, setStore, and deleteStore.
+
 ## [2.8.0](https://github.com/wickeyware/wickeyappstore/compare/2.7.2...2.8.0) - 2018-04-19
 
 ### Fixed